+++ lib/src/widgets/settings_panel/lit_settings_panel_controller.dart
@@ -18,7 +18,14 @@ class LitSettingsPanelController {
   /// on the [StatefulWidget] by overriding its dispose method.
   void dispose() {
     _listeners = [];
-    animationController.dispose();
+    try {
+      animationController.dispose();
+    } catch (e) {
+      print("Failed to dispose the `LitSettingsPanelController`. " +
+          "Have you already disposed the controller on your `StatefulWidget`?" +
+          " When using the `LitScaffold`, keep in mind that most controllers " +
+          "will be disposed automatically.");
+    }
   }
 
   void dismissSettingsPanel() {