+++ lib/src/widgets/snackbars/lit_snackbar_controller.dart
@@ -65,7 +65,15 @@ class LitSnackbarController {
   /// to ensure no [TickerProvider] leak will occur.  Call this method on the
   /// [StatefulWidget] by overriding its dispose method.
   void dispose() {
-    animationController.dispose();
+    _listeners = [];
+    try {
+      animationController.dispose();
+    } catch (e) {
+      print("Failed to dispose the `LitSnackbarController`. " +
+          "Have you already disposed the controller on your `StatefulWidget`?" +
+          " When using the `LitScaffold`, keep in mind that most controllers " +
+          "will be disposed automatically.");
+    }
   }
 
   /// Slides in the [LitSnackbar] by playing the [AnimationController]'s