+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `LitInfoIcon` to `art` library.
 - Add `LitStaticLoadingScreen` to `screens` library.
 - Improve error handling when calling `dispose` on several custom controller classes.
+- Update `LitSansSerifStyles` type system according to `Material Design` Guidelines ([https://material.io/design/typography/the-type-system.html#type-scale](https://material.io/design/typography/the-type-system.html#type-scale)).
 
 ## 1.2.0